moved playlist item into its own folder
