+++ frontend/src/scenes/session-recordings/playlist/SessionRecordingsPlaylist.scss
