+++ frontend/src/scenes/session-recordings/playlist/SessionRecordingsPlaylist.tsx
@@ -1,159 +1,26 @@
 import { useRef } from 'react'
 import { useActions, useValues } from 'kea'
-import { colonDelimitedDuration, range } from '~/lib/utils'
+import { range } from '~/lib/utils'
 import { SessionRecordingType } from '~/types'
 import {
     defaultPageviewPropertyEntityFilter,
     PLAYLIST_LIMIT,
     sessionRecordingsListLogic,
 } from './sessionRecordingsListLogic'
-import { asDisplay } from 'scenes/persons/PersonHeader'
 import './SessionRecordingsPlaylist.scss'
-import { TZLabel } from 'lib/components/TimezoneAware'
-import { SessionRecordingPlayer } from './player/SessionRecordingPlayer'
+import { SessionRecordingPlayer } from '../player/SessionRecordingPlayer'
 import { EmptyMessage } from 'lib/components/EmptyMessage/EmptyMessage'
 import { LemonButton } from '@posthog/lemon-ui'
-import { IconChevronLeft, IconChevronRight, IconSchedule } from 'lib/components/icons'
-import { SessionRecordingsFilters } from './filters/SessionRecordingsFilters'
+import { IconChevronLeft, IconChevronRight } from 'lib/components/icons'
+import { SessionRecordingsFilters } from '../filters/SessionRecordingsFilters'
 import clsx from 'clsx'
-import { Tooltip } from 'lib/components/Tooltip'
 import { LemonSkeleton } from 'lib/components/LemonSkeleton'
 import { LemonTableLoader } from 'lib/components/LemonTable/LemonTableLoader'
-import { PropertyIcon } from 'lib/components/PropertyIcon'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { FEATURE_FLAGS } from 'lib/constants'
+import { SessionRecordingPlaylistItem } from 'scenes/session-recordings/playlist/SessionRecordingsPlaylistItem'
 
 interface SessionRecordingsTableProps {
     personUUID?: string
     isPersonPage?: boolean
-}
-
-interface SessionRecordingPlaylistItemProps {
-    recording: SessionRecordingType
-    isActive: boolean
-    onClick: () => void
-    onPropertyClick: (property: string, value?: string) => void
-}
-
-const SessionRecordingPlaylistItem = ({
-    recording,
-    isActive,
-    onClick,
-    onPropertyClick,
-}: SessionRecordingPlaylistItemProps): JSX.Element => {
-    const formattedDuration = colonDelimitedDuration(recording.recording_duration)
-    const durationParts = formattedDuration.split(':')
-
-    const { featureFlags } = useValues(featureFlagLogic)
-
-    const listIcons = featureFlags[FEATURE_FLAGS.RECORDING_LIST_ICONS] || 'none'
-    const iconClassnames = clsx(
-        'SessionRecordingsPlaylist__list-item__property-icon text-base text-muted-alt',
-        !isActive && 'opacity-75'
-    )
-    const iconPropertyKeys = ['$browser', '$device_type', '$os', '$geoip_country_code']
-    const iconProperties = recording.properties || recording.person?.properties || {}
-
-    const indicatorRight = listIcons === 'bottom' || listIcons === 'none' || listIcons === 'middle' || !listIcons
-
-    const propertyIcons = (
-        <div className="flex flex-row flex-nowrap shrink-0 gap-1">
-            {iconPropertyKeys.map((property) => (
-                <PropertyIcon
-                    key={property}
-                    onClick={onPropertyClick}
-                    className={iconClassnames}
-                    property={property}
-                    value={iconProperties?.[property]}
-                    tooltipTitle={(_, value) => (
-                        <div className="text-center">
-                            Click to filter for
-                            <br />
-                            <span className="font-medium">{value ?? 'N/A'}</span>
-                        </div>
-                    )}
-                />
-            ))}
-        </div>
-    )
-
-    const duration = (
-        <span className="flex items-center font-semibold">
-            <IconSchedule className={iconClassnames} />
-            <span>
-                <span className={clsx(durationParts[0] === '00' && 'opacity-50 font-normal')}>{durationParts[0]}:</span>
-                <span
-                    className={clsx({
-                        'opacity-50 font-normal': durationParts[0] === '00' && durationParts[1] === '00',
-                    })}
-                >
-                    {durationParts[1]}:
-                </span>
-                {durationParts[2]}
-            </span>
-        </span>
-    )
-
-    return (
-        <li
-            key={recording.id}
-            className={clsx(
-                'SessionRecordingsPlaylist__list-item',
-                'flex flex-row py-2 pr-4 pl-0 cursor-pointer relative overflow-hidden',
-                isActive && 'bg-primary-highlight font-semibold',
-                indicatorRight && 'pl-4'
-            )}
-            onClick={() => onClick()}
-        >
-            {!indicatorRight && (
-                <div className="w-2 h-2 mx-2">
-                    {!recording.viewed ? (
-                        <Tooltip title={'Indicates the recording has not been watched yet'}>
-                            <div
-                                className="w-2 h-2 mt-2 rounded-full bg-primary-light"
-                                aria-label="unwatched-recording-label"
-                            />
-                        </Tooltip>
-                    ) : null}
-                </div>
-            )}
-            <div className="grow">
-                <div className="flex items-center justify-between">
-                    <div className="truncate font-medium text-primary ph-no-capture">
-                        {asDisplay(recording.person, 25)}
-                    </div>
-
-                    <div className="flex-1" />
-
-                    {listIcons === 'top-right' && propertyIcons}
-                    {listIcons === 'bottom-right' && duration}
-                    {indicatorRight && !recording.viewed && (
-                        <Tooltip title={'Indicates the recording has not been watched yet'}>
-                            <div
-                                className="w-2 h-2 rounded-full bg-primary-light"
-                                aria-label="unwatched-recording-label"
-                            />
-                        </Tooltip>
-                    )}
-                </div>
-
-                {listIcons === 'middle' && <div>{propertyIcons}</div>}
-
-                <div className="flex items-center justify-between">
-                    <TZLabel
-                        className="overflow-hidden text-ellipsis"
-                        time={recording.start_time}
-                        formatDate="MMMM DD, YYYY"
-                        formatTime="h:mm A"
-                    />
-                    <div className="flex items-center gap-2 flex-1 justify-end">
-                        {listIcons === 'bottom' && propertyIcons}
-                        {listIcons !== 'bottom-right' ? duration : propertyIcons}
-                    </div>
-                </div>
-            </div>
-        </li>
-    )
 }
 
 export function SessionRecordingsPlaylist({ personUUID }: SessionRecordingsTableProps): JSX.Element {
